backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
google-genai>=1.0.0
python-multipart==0.0.12
pydantic==2.10.0
python-dotenv==1.0.1
